+++ crates/core/Cargo.toml
@@ -40,7 +40,7 @@ path-absolutize = { version = "3.1.1", optional = false, features = [
   "use_unix_paths_on_wasm",
 ] }
 getrandom = { version = "0.2.11", optional = true }
-lazy_static = {version = "1.4.0", optional = true }
+lazy_static = { version = "1.4.0", optional = true }
 walkdir = { version = "2.3.3", optional = true }
 fs-err = { version = "2.11.0" }
 
@@ -60,6 +60,7 @@ default = [
   "network_requests",
   "language-parsers",
   "non_wasm",
+  "matches",
 ]
 external_functions_common = []
 external_functions = ["external_functions_common", "dep:marzano-externals"]
@@ -90,3 +91,4 @@ grit-parser = ["marzano-language/grit-parser"]
 absolute_filename = []
 non_wasm = ["absolute_filename"]
 test_utils = []
+matches = []
